Refactor Edge Worker CLI for smaller module
