+++ providers/edge3/tests/unit/edge3/cli/test_worker.py
@@ -90,7 +90,7 @@ class _MockPopen(Popen):
         return self.generated_returncode
 
 
-class TestEdgeWorkerCli:
+class TestEdgeWorker:
     parser: argparse.ArgumentParser
 
     @classmethod
