+++ providers/edge3/tests/unit/edge3/worker_api/routes/test_worker.py
@@ -21,7 +21,7 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from airflow.providers.edge3.cli.edge_command import _EdgeWorkerCli
+from airflow.providers.edge3.cli.worker import EdgeWorker
 from airflow.providers.edge3.models.edge_worker import EdgeWorkerModel, EdgeWorkerState
 from airflow.providers.edge3.worker_api.datamodels import WorkerQueueUpdateBody, WorkerStateBody
 from airflow.providers.edge3.worker_api.routes._v2_compat import HTTPException
@@ -41,8 +41,8 @@ pytestmark = pytest.mark.db_test
 
 class TestWorkerApiRoutes:
     @pytest.fixture
-    def cli_worker(self, tmp_path: Path) -> _EdgeWorkerCli:
-        test_worker = _EdgeWorkerCli(str(tmp_path / "mock.pid"), "mock", None, 8, 5, 5)
+    def cli_worker(self, tmp_path: Path) -> EdgeWorker:
+        test_worker = EdgeWorker(str(tmp_path / "mock.pid"), "mock", None, 8, 5, 5)
         return test_worker
 
     @pytest.fixture(autouse=True)
@@ -77,7 +77,7 @@ class TestWorkerApiRoutes:
             pytest.param(["default", "default2"], id="with-queues"),
         ],
     )
-    def test_register(self, session: Session, input_queues: list[str] | None, cli_worker: _EdgeWorkerCli):
+    def test_register(self, session: Session, input_queues: list[str] | None, cli_worker: EdgeWorker):
         body = WorkerStateBody(
             state=EdgeWorkerState.STARTING,
             jobs_active=0,
@@ -170,7 +170,7 @@ class TestWorkerApiRoutes:
 
         assert redefine_state(worker_state, body_state) == expected_state
 
-    def test_set_state(self, session: Session, cli_worker: _EdgeWorkerCli):
+    def test_set_state(self, session: Session, cli_worker: EdgeWorker):
         queues = ["default", "default2"]
         rwm = EdgeWorkerModel(
             worker_name="test2_worker",
